Dcx: Start DcxFile
